% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-asis.R
\name{as-is}
\alias{as-is}
\alias{vec_ptype2.AsIs}
\title{AsIs S3 class}
\usage{
\method{vec_ptype2}{AsIs}(x, y, ..., x_arg = "", y_arg = "")
}
\description{
These functions help the base AsIs class fit into the vctrs type system
by providing coercion and casting functions.
}
\keyword{internal}
