% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{df_detect_complete}
\alias{df_detect_complete}
\title{Complete}
\usage{
df_detect_complete(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A logical vector with the same size as \code{x}.
}
\description{
\code{df_detect_complete()} detects "complete" rows of a data frame. A
row is only considered complete if all elements of that row are non-missing.
To compare, \code{!vec_equal_na(x)} detects rows that have at least one
non-missing value.
}
\examples{
df <- data_frame(
  x = c(1, 2, NA, 4, NA),
  y = c("a", "b", NA, "d", "e")
)

# This returns `TRUE` where all elements of the row are non-missing.
# Compare that with `!vec_equal_na()`, which detects rows that have at
# least one non-missing value.
df2 <- df
df2$all_non_missing <- df_detect_complete(df)
df2$any_non_missing <- !vec_equal_na(df)
df2
}
\seealso{
\code{\link[stats:complete.cases]{stats::complete.cases()}}
}
