% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.R
\name{dim}
\alias{dim}
\alias{vec_dim}
\alias{vec_dim_n}
\title{Actual vector dimensions}
\usage{
vec_dim(x)

vec_dim_n(x)
}
\arguments{
\item{x}{A vector}
}
\description{
\itemize{
\item \code{vec_dim_n()} gives the dimensionality (i.e. number of dimensions)
\item \code{vec_dim()} returns the size of each dimension
}

These functions access the raw \code{"dim"} attribute of the object
and do not dispatch over the \code{\link[=dim]{dim()}} generic.
}
\details{
Unlike base R, we treat vectors with \code{NULL} dimensions as 1d. This
simplifies the type system by eliding a special case. Compared to the base R
equivalent, \code{vec_dim()} returns \code{\link[=length]{length()}}, not \code{NULL}, when \code{x} is 1d.
}
\examples{
# Compared to base R
x <- 1:5
dim(x)
vec_dim(x)
}
\seealso{
\code{\link[=dim2]{dim2()}}, a variant of \code{\link[=dim]{dim()}} that returns \code{\link[=length]{length()}} if an object
doesn't have dimensions.
}
