% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.R
\name{dim2}
\alias{dim2}
\title{Perceived vector dimensions}
\usage{
dim2(x)
}
\arguments{
\item{x}{A vector}
}
\description{
\code{dim2()} is a variant of \code{\link[=dim]{dim()}} that returns \code{\link[=vec_size]{vec_size()}} if an object
doesn't have dimensions.
}
\details{
Unlike base R, we treat vectors with \code{NULL} dimensions as 1d. This
simplifies the type system by eliding a special case. Compared to the base R
equivalent, \code{dim2()} returns \code{\link[=length]{length()}}, not \code{NULL}, when \code{x} is 1d.
}
\examples{
# Compared to base R
x <- 1:5
dim(x)
vec_dim(x)
}
\seealso{
\code{\link[=vec_dim]{vec_dim()}}, a variant that never dispatches over the \code{\link[=dim]{dim()}} generic.
}
