% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faq.R
\name{faq-error-incompatible-attributes}
\alias{faq-error-incompatible-attributes}
\title{FAQ - Error: Some attributes are incompatible}
\description{
TODO. Tracked in r-lib/vctrs#1016

This error occurs when \code{\link[=vec_ptype2]{vec_ptype2()}} or \code{\link[=vec_cast]{vec_cast()}} are supplied
vectors of the same classes with different attributes.
}
