% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faq.R
\name{faq-warning-convert-inputs}
\alias{faq-warning-convert-inputs}
\title{FAQ - Warning: Convert all inputs to the same class}
\description{
TODO. Tracked in r-lib/vctrs#1016

This warning occurs before falling back to \code{data.frame} when
incompatible data frame classes are supplied.
}
