% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-vctr.R
\docType{data}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{+,ANY,vctrs_vctr-method}
\alias{+,vctrs_vctr,ANY-method}
\alias{+,vctrs_vctr,missing-method}
\alias{c,vctrs_vctr-method}
\alias{cbind2,ANY,vctrs_vctr-method}
\alias{cbind2,vctrs_vctr,ANY-method}
\alias{show,vctrs_vctr-method}
\title{For S4 methods that require a documentation entry but only clutter the index.}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
