% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-integer64.R
\name{vec_ptype_full.integer64}
\alias{vec_ptype_full.integer64}
\alias{vec_ptype_abbr.integer64}
\alias{vec_ptype2.integer64}
\alias{vec_cast.integer64}
\title{64 bit integers}
\usage{
\method{vec_ptype_full}{integer64}(x, ...)

\method{vec_ptype_abbr}{integer64}(x, ...)

\method{vec_ptype2}{integer64}(x, y, ...)

\method{vec_cast}{integer64}(x, to, ...)
}
\description{
A \code{integer64} is a 64 bits integer vector, implemented in the \code{bit64} package.
}
\details{
These functions help the \code{integer64} class from \code{bit64} in to
the vctrs type system by providing coercion functions
and casting functions.
}
\keyword{internal}
