% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-list-of.R
\name{list-of-attributes}
\alias{list-of-attributes}
\alias{list_of_ptype}
\alias{list_of_size}
\title{\code{list_of} attributes}
\usage{
list_of_ptype(x)

list_of_size(x)
}
\arguments{
\item{x}{A \link[=list_of]{list_of}.}
}
\description{
\itemize{
\item \code{list_of_ptype()} returns the \code{ptype} required by the \code{list_of}.
If no \code{ptype} is required, then \code{NULL} is returned.
\item \code{list_of_size()} returns the \code{size} required by the \code{list_of}.
If no \code{size} is required, then \code{NULL} is returned.
}
}
\examples{
x <- list_of(1, 2)
list_of_ptype(x)
list_of_size(x)

x <- list_of(.ptype = integer(), .size = 5)
list_of_ptype(x)
list_of_size(x)
}
