% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact.R
\name{list_compact}
\alias{list_compact}
\title{Drop missing or empty elements}
\usage{
list_compact(x, ..., drop = c("missing", "empty"))
}
\arguments{
\item{x}{A list.}

\item{...}{These dots are for future extensions and must be empty.}

\item{drop}{A string specifying the elements to drop. One of:
\itemize{
\item \code{"missing"}: Drop all \code{NULL} elements.
\item \code{"empty"}: Drop all size 0 elements. This includes \code{NULL} elements along
with empty vectors like \code{integer(0)}.
}}
}
\description{
\code{list_compact()} removes missing or empty elements from a list.
}
\examples{
x <- list(1, NULL, integer(), 2)

# By default, only `NULL` elements are dropped
list_compact(x)

# You can also drop any empty element
list_compact(x, drop = "empty")
}
