% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-data-frame.R
\name{new_data_frame}
\alias{new_data_frame}
\alias{vec_ptype2.data.frame}
\alias{vec_cast.data.frame}
\title{Data frame class}
\usage{
new_data_frame(x = list(), n = NULL, ..., class = character())

\method{vec_ptype2}{data.frame}(x, y, ...)

\method{vec_cast}{data.frame}(x, to, ...)
}
\arguments{
\item{x}{A named list of equal-length vectors. The lengths are not
checked; it is responsibility of the caller to make sure they are
equal.}

\item{n}{Number of rows. If \code{NULL}, will be computed from the length of
the first element of \code{x}.}

\item{..., class}{Additional arguments for creating subclasses.}
}
\description{
A \code{data.frame} \code{\link[=data.frame]{data.frame()}} is a list with "row.names" attribute. Each
element of the list must be named, and of the same length. These functions
help the base data.frame classes fit in to the vctrs type system by
providing constructors, coercion functions, and casting functions.
}
\examples{
new_data_frame(list(x = 1:10, y = 10:1))
}
\keyword{internal}
