% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-date-time.R
\name{new_date}
\alias{new_date}
\alias{new_datetime}
\alias{new_duration}
\alias{vec_type2.Date}
\alias{vec_type2.POSIXt}
\alias{vec_type2.difftime}
\alias{vec_cast.Date}
\alias{vec_cast.POSIXt}
\alias{vec_cast.difftime}
\alias{vec_arith.Date}
\alias{vec_arith.POSIXct}
\alias{vec_arith.difftime}
\title{Date, date-time, and duration S3 classes}
\usage{
new_date(x = double())

new_datetime(x = double(), tzone = "")

new_duration(x = double(), units = c("secs", "mins", "hours", "days",
  "weeks"))

\method{vec_type2}{Date}(x, y)

\method{vec_type2}{POSIXt}(x, y)

\method{vec_type2}{difftime}(x, y)

\method{vec_cast}{Date}(x, to)

\method{vec_cast}{POSIXt}(x, to)

\method{vec_cast}{difftime}(x, to)

\method{vec_arith}{Date}(op, x, y)

\method{vec_arith}{POSIXct}(op, x, y)

\method{vec_arith}{difftime}(op, x, y)
}
\arguments{
\item{x}{A double vector representing the number of days since UNIX
epoch for \code{new_date()}, number of seconds since UNIX epoch for
\code{new_datetime()}, and number of \code{units} for \code{new_duration()}.}

\item{tzone}{Time zone. Either \code{""} for local time zone, or a value
from \code{\link[=OlsonNames]{OlsonNames()}}}

\item{units}{Units of duration.}
}
\description{
These function help the date (\link{Date}), datetime (\link{POSIXct}), and duration
(\link{difftime}) classes fit into the vctrs type system by providing low-level
constructors, as well as coercion and casting generics + methods.
}
\examples{
new_date(0)
new_datetime(0, tzone = "UTC")
new_duration(1, "hour")
}
\keyword{internal}
