% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-integer64.R
\name{new_int64}
\alias{new_int64}
\alias{vec_ptype_full.integer64}
\alias{vec_ptype_abbr.integer64}
\alias{vec_type2.integer64}
\alias{vec_cast.integer64}
\title{64 bit integers}
\usage{
new_int64(x = bit64::integer64(), ..., class = character())

\method{vec_ptype_full}{integer64}(x)

\method{vec_ptype_abbr}{integer64}(x)

\method{vec_type2}{integer64}(x, y)

\method{vec_cast}{integer64}(x, to)
}
\arguments{
\item{x}{64 bit integer vector}

\item{..., class}{Used to for subclasses.}
}
\description{
A \code{integer64} vector is a 64 bits integer vector. Details
are implemented in the \code{bit64} package.
}
\details{
These functions help the \code{integer64} class from \code{bit64} in to
the vctrs type system by providing constructors, coercion functions,
and casting functions. \code{new_int64()} is a low-level
constructor that only checks the type is valid, so
is for expert use only.
}
\keyword{internal}
