% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrd.R
\name{new_rcrd}
\alias{new_rcrd}
\alias{ses}
\alias{rcrd}
\title{rcrd (record) S3 class}
\usage{
new_rcrd(fields, ..., class = character())
}
\arguments{
\item{fields}{A list. It must possess the following properties:
\itemize{
\item no attributes (apart from names)
\item syntactic names
\item length 1 or greater
\item elements are vectors
\item elements have equal length
}}

\item{...}{Additional attributes}

\item{class}{Name of subclass.}
}
\description{
The rcrd class extends \link{vctr}. A rcrd is composed of 1 or more \link{field}s,
which must be vectors of the same length. Is designed specifically for
classes that can naturally be decomposed into multiple vectors of the same
length, like \link{POSIXlt}, but where the organisation should be considered
an implementation detail invisible to the user (unlike a \link{data.frame}).
}
\keyword{internal}
