% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{new_record}
\alias{new_record}
\alias{ses}
\alias{record}
\title{record S3 class}
\usage{
new_record(fields, ..., class = character())
}
\arguments{
\item{fields}{A list. It must possess the following properties:
\itemize{
\item no attributes (apart from names)
\item syntactic names
\item length 1 or greater
\item elements are vectors
\item elements have equal length
}}

\item{...}{Additional attributes}

\item{class}{Name of subclass.}
}
\description{
The record class extends \link{vctr}. A record is composed of 1 or more \link{field}s,
which must be vectors of the same length. Is designed specifically for
classes that can naturally be decomposed into multiple vectors of the same
length, like \link{POSIXlt}, but where the organisation should be considered
an implementation detail invisible to the user (unlike a \link{data.frame}).
}
\keyword{internal}
