% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{obj_hash}
\alias{obj_hash}
\title{Hash an object}
\usage{
obj_hash(x, pool = TRUE)
}
\arguments{
\item{x}{A vector}

\item{pool}{A logical. Should R's global string pool be used to hash
characters? If \code{TRUE}, the hashing will be faster, but will not be
reproducible between R sessions.}
}
\value{
The hash value as a raw vector of size 4.
}
\description{
\Sexpr[results=rd, stage=render]{vctrs:::lifecycle("experimental")}

\code{obj_hash()} hashes an object by coding each element of \code{x} with 32-bit
hashes and combining them together. The hash is returned as a raw vector of
size 4.
}
\details{
\code{obj_hash()} is extremely experimental, and no strong reproducibility
assumptions should be made. Because of this, if you use \code{obj_hash()} the
hash results should not be tested in unit tests as they might change without
warning.

It is possible that some objects have a non-trivial chance to produce the
same hash value.
}
\examples{
obj_hash(1:2)
obj_hash(mtcars)
}
\keyword{internal}
