% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-str.R
\name{obj_print}
\alias{obj_print}
\alias{obj_print_header}
\alias{obj_print_data}
\alias{obj_print_footer}
\alias{obj_str}
\alias{obj_str_header}
\alias{obj_str_data}
\alias{obj_str_footer}
\title{\code{print()} and \code{str()} generics.}
\usage{
obj_print(x, ...)

obj_print_header(x, ...)

obj_print_data(x, ...)

obj_print_footer(x, ...)

obj_str(x, ...)

obj_str_header(x, ...)

obj_str_data(x, ...)

obj_str_footer(x, ...)
}
\arguments{
\item{x}{A vector}

\item{...}{Additional arguments passed on to methods. See \code{\link[=print]{print()}} and
\code{\link[=str]{str()}} for commonly used options}
}
\description{
These are constructed to be more easily extensible since you can override
the \verb{_header()}, \verb{_data()} or \verb{_footer()} components individually. The
default methods are built on top of \code{format()}.
}
\keyword{internal}
