% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parallel-operators}
\alias{parallel-operators}
\alias{vec_pany}
\alias{vec_pall}
\title{Parallel \code{any()} and \code{all()}}
\usage{
vec_pany(..., .na_rm = FALSE, .size = NULL)

vec_pall(..., .na_rm = FALSE, .size = NULL)
}
\arguments{
\item{...}{Logical vectors. These will be \link[=vector_recycling_rules]{recycled}
to their common size.}

\item{.na_rm}{Should missing values be removed?}

\item{.size}{An optional output size that overrides the common size of the
inputs in \code{...}.}
}
\description{
These functions are variants of \code{\link[=any]{any()}} and \code{\link[=all]{all()}} that work in parallel on
multiple inputs at once. They work similarly to how \code{\link[=pmin]{pmin()}} and \code{\link[=pmax]{pmax()}} are
parallel variants of \code{\link[=min]{min()}} and \code{\link[=max]{max()}}.
}
\details{
\code{vec_pany()} and \code{vec_pall()} are consistent with \code{\link[=any]{any()}} and \code{\link[=all]{all()}} when
there are no inputs to process in parallel:
\itemize{
\item \code{any()} returns \code{FALSE} with no inputs. Similarly, \code{vec_pany(.size = 1)}
and \code{vec_pany(NA, .na_rm = TRUE)} both return \code{FALSE}.
\item \code{all()} returns \code{TRUE} with no inputs. Similarly, \code{vec_pall(.size = 1)}
and \code{vec_pall(NA, .na_rm = TRUE)} both return \code{TRUE}.
}
}
\examples{
x <- c(TRUE, FALSE, NA, TRUE, NA)
y <- c(FALSE, FALSE, TRUE, TRUE, NA)

vec_pany(x, y)
vec_pall(x, y)

# Missing values can be removed from the computation
vec_pany(x, y, .na_rm = TRUE)
vec_pall(x, y, .na_rm = TRUE)

# Same empty behavior as `any()` and `all()`
vec_pany(.size = 1)
any()

vec_pall(.size = 1)
all()
}
