% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated.R
\name{repeated}
\alias{repeated}
\alias{as_repeated}
\alias{new_repeated}
\alias{is_repeated}
\title{Construct "repeated" objects}
\usage{
repeated(..., .strict = TRUE, .type = NULL)

as_repeated(x, ...)

new_repeated(x, type)

is_repeated(x)
}
\arguments{
\item{...}{Vectors to coerce. All vectors must be 1d (i.e. no data
frames, matrices or arrays).}

\item{.strict}{If \code{.strict = FALSE}, there will always be a common type for
any pair of vectors. This will be a character vector for factors with
different level sets, and a list for everything else.}
}
\description{
A repeated object is a list where each element has the same type.
Modifying the list with \code{$}, \code{[}, and \code{[[} preserves the constraint
by coercing all input items.
}
\examples{
x <- repeated(1:3, 5:6, 10:15)
tibble::tibble(x = x)

vec_c(repeated(1, 2), repeated(FALSE, TRUE))
}
