% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{shape}
\alias{shape}
\alias{new_shape}
\alias{as_shape}
\title{Create a shape}
\usage{
shape(..., data.frame = FALSE)

new_shape(x, data.frame = FALSE)

as_shape(x)
}
\arguments{
\item{data.frame}{Does this shape represent a data frame? Data frames
must be 2 dimensional, and can only be recycled along the first dimension.}

\item{x}{For \code{shape()}, a vector giving dimensions of an object.
For \code{as_shape()}, an existing data vector}
}
\description{
\code{new_shape()} is the low-level constructor, \code{shape()} is a helper for
interactive exploration, \code{as_shape()} computes the shape of an existing
object. Use \code{max()} to compute the "maximum" shape of a set of shapes,
obeying the recycling rules.
}
\examples{
shape(1, 2, 3)
as_shape(mtcars)

max(shape(1), shape(10))
\dontrun{
max(shape(2), shape(10))
}

max(shape(10), shape(0))
max(shape(10), shape(10, 2))
max(shape(1, 5, 1), shape(2, 5, 10))
}
\keyword{internal}
