% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{stop_index_oob_position}
\alias{stop_index_oob_position}
\alias{stop_index_oob_name}
\title{Out-of-bounds errors}
\usage{
stop_index_oob_position(i, size, ..., .subclass = NULL)

stop_index_oob_name(i, names, ..., .subclass = NULL)
}
\arguments{
\item{i}{For \code{stop_index_oob_position()}, a numeric vector of
positions. For \code{stop_index_oob_name()}, a character vector of
names. \code{i} may contain both out-of-bounds and within-bounds
elements, only the former are used to construct the error
message.}

\item{size}{The length of the vector to subset from.}

\item{...}{Additional data to be stored in the condition object.}

\item{.subclass}{Subclass of the condition. This allows your users
to selectively handle the conditions signalled by your functions.}

\item{names}{The names of the vector to subset from.}
}
\description{
\itemize{
\item \code{stop_index_oob_position()} throws errors of class
\code{vctrs_error_index_oob_position} containing fields \code{i} and
\code{size}.
\item \code{stop_index_oob_name()} throws errors of class
\code{vctrs_error_index_oob_name} containing fields \code{i} and \code{names}.
}
}
