% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{stop_subscript_oob_location}
\alias{stop_subscript_oob_location}
\alias{stop_subscript_oob_name}
\title{Out-of-bounds errors}
\usage{
stop_subscript_oob_location(i, size, ..., .subclass = NULL)

stop_subscript_oob_name(i, names, ..., .subclass = NULL)
}
\arguments{
\item{i}{For \code{stop_subscript_oob_location()}, a numeric vector of
positions. For \code{stop_subscript_oob_name()}, a character vector of
names. \code{i} may contain both out-of-bounds and within-bounds
elements, only the former are used to construct the error
message.}

\item{size}{The length of the vector to subset from.}

\item{...}{Additional data to be stored in the condition object.}

\item{.subclass}{This argument was renamed to \code{class} in rlang
0.4.2.  It will be deprecated in the next major version. This is
for consistency with our conventions for class constructors
documented in \url{https://adv-r.hadley.nz/s3.html#s3-subclassing}.}

\item{names}{The names of the vector to subset from.}
}
\description{
\itemize{
\item \code{stop_subscript_oob_location()} throws errors of class
\code{vctrs_error_subscript_oob_location} containing fields \code{i} and
\code{size}.
\item \code{stop_subscript_oob_name()} throws errors of class
\code{vctrs_error_subscript_oob_name} containing fields \code{i} and \code{names}.
}
}
