% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-tibble.R
\name{tbl_cast.tbl_df}
\alias{tbl_cast.tbl_df}
\title{Double dispatch methods for tibbles}
\usage{
\method{tbl_cast}{tbl_df}(x, to, ..., x_arg = "x", to_arg = "to")
}
\arguments{
\item{x}{A tabular vector.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{These dots are for future extensions and must be empty.}

\item{x_arg}{Argument names for \code{x} and \code{y}. These are used
in error messages to inform the user about the locations of
incompatible types (see \code{\link[=stop_incompatible_type]{stop_incompatible_type()}}).}

\item{to_arg}{Argument names for \code{x} and \code{to}. These are used
in error messages to inform the user about the locations of
incompatible types (see \code{\link[=stop_incompatible_type]{stop_incompatible_type()}}).}
}
\description{
Double dispatch methods for tibbles
}
