% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknown.R
\name{unknown}
\alias{unknown}
\alias{vec_type2.unknown}
\title{A 1d vector of unknown type}
\usage{
unknown(n = 0)

\method{vec_type2}{unknown}(x, y)
}
\arguments{
\item{n}{Length of vector}
}
\description{
This is a \href{new_partial}{partial type} used to represents logical vectors
that only contain \code{NA}. These require special handling because we want to
allow \code{NA} to specify missingness without requiring a type.
}
\examples{
vec_ptype()
vec_ptype(NA)

vec_c(NA, factor("x"))
vec_c(NA, Sys.Date())
vec_c(NA, Sys.time())
vec_c(NA, list(1:3, 4:5))
}
\keyword{internal}
