% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unspecified.R
\name{unspecified}
\alias{unspecified}
\alias{vec_unspecified_cast}
\alias{vec_type2.vctrs_unspecified}
\title{A 1d vector of unspecified type}
\usage{
unspecified(n = 0)

vec_unspecified_cast(x, to)

\method{vec_type2}{vctrs_unspecified}(x, y)
}
\arguments{
\item{n}{Length of vector}
}
\description{
This is a \href{new_partial}{partial type} used to represent logical vectors
that only contain \code{NA}. These require special handling because we want to
allow \code{NA} to specify missingness without requiring a type.
}
\details{
\code{vec_unspecified_cast()} is a helper to use in your \code{\link[=vec_cast]{vec_cast()}} methods.
See \code{vignette("s3-vector")} for detail.
}
\examples{
vec_ptype()
vec_ptype(NA)

vec_c(NA, factor("x"))
vec_c(NA, Sys.Date())
vec_c(NA, Sys.time())
vec_c(NA, list(1:3, 4:5))
}
\keyword{internal}
