% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-data-frame.R
\name{vctrs-data-frame}
\alias{vctrs-data-frame}
\alias{vec_ptype2.data.frame}
\alias{vec_cast.data.frame}
\title{vctrs methods for data frames}
\usage{
\method{vec_ptype2}{data.frame}(x, y, ...)

\method{vec_cast}{data.frame}(x, to, ...)
}
\description{
These functions help the base data.frame class fit into the vctrs type system
by providing coercion and casting functions.
}
\keyword{internal}
