% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-package.R
\docType{package}
\name{vctrs-package}
\alias{vctrs}
\alias{vctrs-package}
\title{vctrs: Vector Helpers}
\description{
Functions that help you work with vectors, particularly numeric.
Provides a full set of summary, cumulative, rolling, and vectorised
functions for both R and C.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hadley/vctrs}
  \item Report bugs at \url{https://github.com/hadley/vctrs/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
