% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-ts.R
\name{vctrs-ts}
\alias{vctrs-ts}
\title{\code{ts} S3 class}
\description{
These functions help the base \code{ts} class fit into the vctrs type system
by providing coercion and casting functions.
}
\details{
The \code{ts} class is a bit strange for vctrs:
\itemize{
\item It allows for arbitrary storage types.
\item It has a data dependent \code{tsp} attribute.
\item The \code{ts()} function prevents you from creating a size 0 version of a \code{ts}
object, presumably because you can't make the \code{tsp} attribute without some
data.
\item The \verb{stats::[.ts} method drops off all attributes.
\item The \code{c()} default method drops off all attributes.
}

Because of all of this, we aggressively drive coercion towards the underlying
atomic type, which mostly matches the \code{[} and \code{c()} methods.
}
\keyword{internal}
