% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{vec-recode-and-replace}
\alias{vec-recode-and-replace}
\alias{vec_recode_values}
\alias{vec_replace_values}
\title{Recode and replace values}
\usage{
vec_recode_values(
  x,
  ...,
  from,
  to,
  default = NULL,
  multiple_from = FALSE,
  multiple_to = FALSE,
  x_arg = "x",
  from_arg = "from",
  to_arg = "to",
  default_arg = "default",
  ptype = NULL,
  call = current_env()
)

vec_replace_values(
  x,
  ...,
  from,
  to,
  multiple_from = FALSE,
  multiple_to = FALSE,
  x_arg = "x",
  from_arg = "from",
  to_arg = "to",
  call = current_env()
)
}
\arguments{
\item{x}{A vector.}

\item{...}{These dots are for future extensions and must be empty.}

\item{from}{Values present in \code{x} that are mapped to values in \code{to}.

Values present in \code{from} but not in \code{x} are ignored.

Values present in \code{x} but not in \code{from} are recoded to \code{default} when using
\code{vec_recode_values()}, and will be left unchanged when using
\code{vec_replace_values()}.

\code{from} will be \link[=theory-faq-coercion]{cast} to the type of \code{x}.
\itemize{
\item If \code{multiple_from = FALSE}, \code{from} must be a single vector of any size,
which will be \link[=theory-faq-coercion]{cast} to the type of \code{x}.
\item If \code{multiple_from = TRUE}, \code{from} must be a list of vectors of any size,
which will individually be \link[=theory-faq-coercion]{cast} to the type
of \code{x}.
}}

\item{to}{Values to map \code{from} to.

The common type of \code{to} and \code{default} will determine the type of the
output, unless \code{ptype} is provided.
\itemize{
\item If \code{multiple_from = FALSE}, \code{to} must be a single vector of size 1 or the
same size as \code{from}.
\item If \code{multiple_to = TRUE}, \code{to} must be a list of vectors. The list itself
must be size 1 or the same size as \code{from}. Each individual vector in the
list must be size 1 or the same size as \code{x}.
}}

\item{default}{Default value to use when there is a value present in \code{x}
that isn't matched by a value in \code{from}.

By default, a missing value is used as the default value.

If supplied, \code{default} must be the same size as \code{x} or size 1.}

\item{x_arg, from_arg, to_arg, default_arg}{Argument names used in error
messages.}

\item{ptype}{An optional override for the output type, which is usually
computed as the common type of \code{to} and \code{default}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A vector the same size \code{x}.
\itemize{
\item For \code{vec_recode_values()}, the type of the output is computed as the common
type of \code{to} and \code{default}, unless overriden by \code{ptype}.
\item For \code{vec_replace_values()}, the type of the output will have the same type
as \code{x}.
}
}
\description{
\itemize{
\item \code{vec_recode_values()} constructs an entirely new vector by recoding the
values from \code{x} specified in \code{from} to the corresponding values in \code{to}. If
there are values in \code{x} not matched by \code{from}, then they are recoded to the
\code{default} value.
\item \code{vec_replace_values()} updates an existing vector by replacing the values
from \code{x} specified in \code{from} with the corresponding values in \code{to}. In this
case, \code{to} must have the same type as \code{x} and values in \code{x} not matched by
\code{from} pass through untouched.
}
}
\examples{
x <- c(1, 2, 3, 1, 2, 4, NA, 5)

# Imagine you have a pre-existing lookup table
likert <- data.frame(
  from = c(1, 2, 3, 4, 5),
  to = c(
    "Strongly disagree",
    "Disagree",
    "Neutral",
    "Agree",
    "Strongly agree"
  )
)
vec_recode_values(x, from = likert$from, to = likert$to)

# If you don't map all of the values, a `default` is used
x <- c(1, 2, 3, 1, 2, 4, NA, 5, 6, 7)
vec_recode_values(x, from = likert$from, to = likert$to)
vec_recode_values(x, from = likert$from, to = likert$to, default = "Unknown")

# If you want existing `NA`s to pass through, include a mapping for `NA` in
# your lookup table
likert <- data.frame(
  from = c(1, 2, 3, 4, 5, NA),
  to = c(
    "Strongly disagree",
    "Disagree",
    "Neutral",
    "Agree",
    "Strongly agree",
    NA
  )
)
vec_recode_values(x, from = likert$from, to = likert$to, default = "Unknown")

if (require("tibble")) {
  # If you want to partially update `x`, retaining the type of `x` and
  # leaving values not covered by `from` alone, use `vec_replace_values()`
  universities <- c(
    "Duke",
    "Fake U",
    "Duke U",
    NA,
    "Chapel Hill",
    "UNC",
    NA,
    "Duke"
  )

  standardize <- tibble::tribble(
    ~from,         ~to,
    "Duke",        "Duke University",
    "Duke U",      "Duke University",
    "UNC",         "UNC Chapel Hill",
    "Chapel Hill", "UNC Chapel Hill",
  )
  vec_replace_values(
    universities,
    from = standardize$from,
    to = standardize$to
  )

  # In this case, you can use a more powerful feature of
  # `vec_replace_values()`, `multiple_from`, which allows you to provide a
  # list of `from` vectors that each match multiple `from` values to a single
  # `to` value. `tribble()` can help you create these!
  standardize <- tibble::tribble(
    ~from, ~to,
    c("Duke", "Duke U"), "Duke University",
    c("UNC", "Chapel Hill"), "UNC Chapel Hill",
  )

  # Note how `from` is a list column
  standardize

  vec_replace_values(
    universities,
    from = standardize$from,
    to = standardize$to,
    multiple_from = TRUE
  )

  # `vec_replace_values()` is also a useful way to map from or to `NA`
  vec_replace_values(universities, from = NA, to = "Unknown")
  vec_replace_values(universities, from = "Fake U", to = NA)
}
}
