% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{vec_assert}
\alias{vec_assert}
\title{Assert an argument has known prototype and/or size}
\usage{
vec_assert(x, ptype = NULL, size = NULL)
}
\arguments{
\item{x}{A vector argument to check.}

\item{ptype}{Prototype to compare against.}

\item{size}{Size to compare against}
}
\value{
Either an error of class \code{"error_assert"}, or \code{x}, invisibly.
}
\description{
If the prototype doesn't match, an error of class
\code{"error_assert_ptype"} is raised.
If the prototype doesn't match, an error of class
\code{"error_assert_size"} is raised.
Both errors inherit from \code{"error_assert"}.
}
