% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-tibble.R
\name{vec_cast.grouped_df}
\alias{vec_cast.grouped_df}
\title{Double dispatch methods for grouped data frames}
\usage{
\method{vec_cast}{grouped_df}(x, to, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()} and \code{vec_restore()}, these dots are only for future
extensions and should be empty.}
}
\description{
Double dispatch methods for grouped data frames
}
