% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{vec_deduplicate}
\alias{vec_deduplicate}
\title{Calculate the indices necessary to deduplicate and reduplicate a vector.}
\usage{
vec_deduplicate(x)
}
\arguments{
\item{x}{A vector (including a data frame).}
}
\value{
A list containing a vector of the unique indices (like \code{vec_unique_loc(x)}) and the mapping of unique
values in the original vector (like \code{match(x, unique(x))}).
}
\description{
The indices can be used to deduplicate a vector, call a function on the unique values, and then reduplicate.
The result is the same as calling \code{vec_unique_loc(x)} and \code{match(x, unique(x))} but more efficient.
}
\examples{
with_deduplication <- function(f) {
  function(x, ...) {
    dedup_indices <- vec_deduplicate(x)
    f(dedup_indices$unique_loc, ...)[dedup_indices$match_unique_loc]
  }
}

x <- as.character(sample.int(1e5, 10)) |>
  rep(1e6)

x |> with_deduplication(tolower)()

}
