% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-deprecated.R
\name{vec_duplicate_any}
\alias{vec_duplicate_any}
\title{Detect if any values are duplicates}
\usage{
vec_duplicate_any(x)
}
\arguments{
\item{x}{A vector (including a data frame).}
}
\description{
\Sexpr[results=rd, stage=render]{vctrs:::lifecycle("soft-deprecated")}

\code{vec_duplicate_any()} has been renamed to \code{\link[=vec_any_duplicate]{vec_any_duplicate()}} and is
soft-deprecated as of vctrs 0.3.0.
}
\keyword{internal}
