% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-deprecated.R
\name{vec_duplicate_detect}
\alias{vec_duplicate_detect}
\title{Detect duplicate values}
\usage{
vec_duplicate_detect(x)
}
\arguments{
\item{x}{A vector (including a data frame).}
}
\description{
\Sexpr[results=rd, stage=render]{vctrs:::lifecycle("soft-deprecated")}

\code{vec_duplicate_detect()} has been renamed to \code{\link[=vec_duplicate_flg]{vec_duplicate_flg()}} and is
soft-deprecated as of vctrs 0.2.2.
}
\keyword{internal}
