% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-deprecated.R
\name{vec_duplicate_id}
\alias{vec_duplicate_id}
\title{Locate first occurrences}
\usage{
vec_duplicate_id(x)
}
\arguments{
\item{x}{A vector (including a data frame).}
}
\description{
\Sexpr[results=rd, stage=render]{vctrs:::lifecycle("soft-deprecated")}

\code{vec_duplicate_id()} has been renamed to \code{\link[=vec_first_loc]{vec_first_loc()}} and is
soft-deprecated as of vctrs 0.2.2.
}
\keyword{internal}
