% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-deprecated.R
\name{vec_equal_na}
\alias{vec_equal_na}
\title{Missing values}
\usage{
vec_equal_na(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A logical vector the same size as \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{vec_equal_na()} has been renamed to \code{\link[=vec_detect_missing]{vec_detect_missing()}} and is deprecated
as of vctrs 0.5.0.
}
\keyword{internal}
