% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{vec_fill}
\alias{vec_fill}
\title{Fill in missing values with the previous or following value}
\usage{
vec_fill(x, direction = c("down", "up"), leading = FALSE, max_gap = NULL)
}
\arguments{
\item{x}{A vector}

\item{direction}{Direction in which to fill missing values. Must be either
\code{"down"} or \code{"up"}.}

\item{leading}{A logical specifying if leading missing values should be
back-filled with the first non-missing value found in the chosen
\code{direction}.}

\item{max_gap}{A single positive integer specifying the maximum size of the
gap of sequential missing values that will be filled. If \code{NULL}, there is
no limit to the gap size.}
}
\description{
\code{vec_fill()} fills gaps of missing values with the previous or following
non-missing value.
}
\keyword{internal}
