% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{vec_fill_missing}
\alias{vec_fill_missing}
\title{Fill in missing values with the previous or following value}
\usage{
vec_fill_missing(
  x,
  direction = c("down", "up", "downup", "updown"),
  max_gap = NULL
)
}
\arguments{
\item{x}{A vector}

\item{direction}{Direction in which to fill missing values. Must be either
\code{"down"}, \code{"up"}, \code{"downup"}, or \code{"updown"}.}

\item{max_gap}{A single positive integer specifying the maximum size of the
gap of sequential missing values that will be filled. If \code{NULL}, there is
no limit to the gap size.}
}
\description{
\code{vec_fill_missing()} fills gaps of missing values with the previous or
following non-missing value.
}
\keyword{internal}
