% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{vec_first_loc}
\alias{vec_first_loc}
\title{Locate first occurrences}
\usage{
vec_first_loc(x)
}
\arguments{
\item{x}{A vector (including a data frame).}
}
\value{
An integer vector the same size as \code{x}.
}
\description{
\code{vec_first_loc()} returns an integer vector giving, for each element of \code{x},
the location of its first occurrence.
}
\details{
While \code{vec_unique_loc()} and \code{vec_duplicate_loc()} return \emph{only} the
locations of the unique and duplicate values respectively, \code{vec_first_loc()}
returns a vector the same size as \code{x} with locations corresponding to the
first occurrence of each value.
}
\section{Missing values}{

In most cases, missing values are not considered to be equal, i.e.
\code{NA == NA} is not \code{TRUE}. This behaviour would be unappealing here,
so these functions consider all \code{NAs} to be equal. (Similarly,
all \code{NaN} are also considered to be equal.)
}

\examples{
x <- c(10, 10, 20, 30, 30, 40)

vec_first_loc(x)

# Compare with `vec_unique_loc()`, which returns locations
# for only the unique values
vec_unique_loc(x)

# And `vec_duplicate_loc()`, which returns locations for only the
# duplicate values
vec_duplicate_loc(x)

vec_first_loc(c(NA, NA, NaN, NaN))
}
