% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{vec_group}
\alias{vec_group}
\alias{vec_group_id}
\alias{vec_group_rle}
\title{Identify groups}
\usage{
vec_group_id(x)

vec_group_rle(x)
}
\arguments{
\item{x}{A vector}
}
\value{
\itemize{
\item \code{vec_group_id()}: An integer vector with the same size as \code{x}.
\item \code{vec_group_rle()}: A list with two integer vector elements: \code{group} and
\code{length}.
}
}
\description{
\Sexpr[results=rd, stage=render]{vctrs:::lifecycle("experimental")}
\itemize{
\item \code{vec_group_id()} returns an identifier for the group that each element of
\code{x} falls in, constructed in the order that they appear. The number of
groups is also returned as an attribute, \code{n}.
\item \code{vec_group_rle()} locates groups in \code{x} and returns them run length
encoded in the order that they appear. The return value is a list holding
the \code{group} identifiers and the run \code{length} of the corresponding group.
The number of groups is also returned as an attribute, \code{n}.
}
}
\examples{
purrr <- c("p", "u", "r", "r", "r")
vec_group_id(purrr)
vec_group_rle(purrr)

groups <- mtcars[c("vs", "am")]
vec_group_id(groups)
vec_group_rle(groups)

# `vec_group_id()` is equivalent to
vec_match(groups, vec_unique(groups))
}
