% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{vec_group_pos}
\alias{vec_group_pos}
\title{Locate unique groups in a vector}
\usage{
vec_group_pos(x)
}
\arguments{
\item{x}{A vector to locate unique groups for.}
}
\value{
A data frame with two columns and size equal to
\code{vec_size(vec_unique(x))}. The \code{key} column has the same type as \code{x}, and
the \code{id} column is a \code{list_of<integer>}.

Note for complex types, the default \code{data.frame} print method will be
suboptimal, and you will want to coerce into a tibble to better understand
the output.
}
\description{
This locates unique groups in \code{x} and returns both the unique values and
the locations of every appearance of each value. It is used to power
\code{\link[=vec_split]{vec_split()}}.
}
\examples{
vec_group_pos(mtcars$vs)
vec_group_pos(mtcars[c("vs", "am")])

if (require("tibble")) {
  as_tibble(vec_group_pos(mtcars[c("vs", "am")]))
}
}
\seealso{
\link{vec_split}
}
