% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{vec_hash}
\alias{vec_hash}
\title{Hash a vector}
\usage{
vec_hash(x, pool = TRUE)
}
\arguments{
\item{x}{A vector}

\item{pool}{A logical. Should R's global string pool be used to hash
characters? If \code{TRUE}, the hashing will be faster, but will not be
reproducible between R sessions.}
}
\value{
A raw vector of hash values of size \code{4 * vec_size(x)}.
}
\description{
\code{vec_hash()} hashes a vector and returns the hash as a raw vector. Each
element of \code{x} is coded with 32-bit hashes. This results in a raw vector
that is 4 times the size of \code{x}.
}
\examples{
vec_hash(1:2)

vec_hash(mtcars)

# To consistently get a size 4 hash for any
# vector, wrap it in a list
vec_hash(list(mtcars))
}
