% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{vec_init}
\alias{vec_init}
\title{Create a missing vector}
\usage{
vec_init(x, n = 1L)
}
\arguments{
\item{x}{Template of missing vector}

\item{n}{Desired size of result}
}
\description{
Create a missing vector
}
\examples{
vec_init(1:10, 3)
vec_init(Sys.Date(), 5)
vec_init(mtcars, 2)
}
