% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast-list.R
\name{vec_list_cast}
\alias{vec_list_cast}
\title{Cast a list to vector of specific type}
\usage{
vec_list_cast(x, to, ..., x_arg = "", to_arg = "")
}
\arguments{
\item{x}{A list}

\item{to}{Type to coerce to}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
This is a function for developers to use when extending vctrs. It casts
a list to a more specific vectoring type, keeping the length constant.
It does this by discarding (with a warning), any elements after the 1.
It is called from \code{vec_cast.XYZ.list()} methods to preserve symmetry with
\code{vec_cast.list.XYZ()}.
}
\details{
See \code{vignette("s3-vector")} for details.
}
\keyword{internal}
