% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{vec_na}
\alias{vec_na}
\title{Create a missing vector}
\usage{
vec_na(x, n = 1L)
}
\arguments{
\item{x}{Template of missing vector}

\item{n}{Desired size of result}
}
\description{
Create a missing vector
}
\examples{
vec_na(1:10, 3)
vec_na(Sys.Date(), 5)
vec_na(mtcars, 2)
}
