% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{vec_obs}
\alias{vec_obs}
\title{Number of observations}
\usage{
vec_obs(x)
}
\arguments{
\item{x}{A vector}
}
\description{
\code{vec_obs()} returns the number of observations in an object. This is
the length of a 1d vector, or the number of rows in a matrix or data frame.
}
\examples{
vec_obs(1:100)
vec_obs(mtcars)
vec_obs(array(dim = c(3, 5, 10)))
}
