% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{vec_order}
\alias{vec_order}
\alias{vec_order_loc}
\alias{vec_sort}
\title{Order and sort vectors}
\usage{
vec_order(x, direction = "asc", na_value = "largest")

vec_order_loc(x, direction = "asc", na_value = "largest")

vec_sort(x, direction = "asc", na_value = "largest")
}
\arguments{
\item{x}{A vector}

\item{direction}{Direction to sort in.
\itemize{
\item Unless \code{x} is a data frame, this should be a single \code{"asc"} or
\code{"desc"} for ascending or descending order respectively.
\item For data frames, this is allowed to be a vector of \code{"asc"} and \code{"desc"}
with length equal to the number of columns in \code{x} specifying the
direction to order each column in.
}}

\item{na_value}{Treatment of \code{NA} values. \code{NaN} values are treated as
equivalent to \code{NA} values.
\itemize{
\item Unless \code{x} is a data frame, this should be a single \code{"largest"} or
\code{"smallest"} for treating \code{NA} values as the largest or smallest values
respectively.
\item For data frames, this is allowed to be a vector of \code{"largest"} and
\code{"smallest"} with length equal to the number of columns in \code{x} specifying
how \code{NA}s should be treated in each column.
}}
}
\value{
\itemize{
\item \code{vec_order()} an integer vector the same size as \code{x}.
\item \code{vec_sort()} a vector with the same size and type as \code{x}.
\item \code{vec_order_loc()}: A two column data frame with size equal to
\code{vec_size(vec_unique(x))}.
\itemize{
\item A \code{key} column of type \code{vec_ptype(x)}.
\item A \code{loc} column of type list, with elements of type integer.
}
}
}
\description{
\code{vec_order()} computes the order of \code{x}. For data frames, the order is
computed along the rows by computing the order of the first column and
using subsequent columns to break ties.

\code{vec_sort()} sorts \code{x} by computing its order and using \code{vec_slice()} to
rearrange.

\code{vec_order_loc()} returns a data frame containing a \code{key} column with
sorted unique groups, and a \code{loc} column with the locations of each
group in \code{x}. It is similar to \code{\link[=vec_group_loc]{vec_group_loc()}}, except the groups are
returned sorted rather than by first appearance.
}
\details{
Character vectors are ordered in the C-locale. This is different from
\code{base::order()}, which respects \code{base::Sys.setlocale()}, but should
produce more reproducible results between different sessions and platforms.
One immediate difference is that ordering is case-sensitive in the C-locale.
Sorting \code{c("b", "C", "a")} with \code{vec_sort()} will return \code{c("C", "a", "b")},
but with \code{base::order()} will return \code{c("a", "b", "C")} unless
\code{base::order(method = "radix")} is explicitly set, which also uses the
C-locale. That said, this typically does not introduce any practical issues.

Character vectors are always sorted in UTF-8. If any non-UTF-8
characters are detected, they are translated to UTF-8 first.
}
\section{Dependencies of \code{vec_order()}}{

\itemize{
\item \code{\link[=vec_proxy_compare]{vec_proxy_compare()}}
}
}

\section{Dependencies of \code{vec_sort()}}{

\itemize{
\item \code{\link[=vec_proxy_compare]{vec_proxy_compare()}}
\item \code{\link[=vec_order]{vec_order()}}
\item \code{\link[=vec_slice]{vec_slice()}}
}
}

\examples{
x <- round(sample(runif(5), 9, replace = TRUE), 3)
x <- c(x, NA)

vec_order(x)
vec_sort(x)
vec_sort(x, "desc")

# Can also handle data frames
df <- data.frame(g = sample(2, 10, replace = TRUE), x = x)
vec_order(df)
vec_sort(df)
vec_sort(df, "desc")

# For data frames, `direction` and `na_value` are allowed to be vectors
# with length equal to the number of columns in the data frame
vec_sort(
  df,
  direction = c("desc", "asc"),
  na_value = c("largest", "smallest")
)

# `vec_order_loc()` is similar to `vec_group_loc()`, except keys are
# returned ordered rather than by first appearance.
vec_order_loc(df)
vec_group_loc(df)
}
