% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{vec_sort}
\alias{vec_sort}
\alias{vec_order_base}
\title{Order vectors}
\usage{
vec_sort(x, direction = c("asc", "desc"), na_value = c("largest", "smallest"))

vec_order_base(
  x,
  direction = c("asc", "desc"),
  na_value = c("largest", "smallest")
)
}
\arguments{
\item{x}{A vector}

\item{direction}{Direction to sort in. Defaults to \code{asc}ending.}

\item{na_value}{Should \code{NA}s be treated as the largest or smallest values?}
}
\value{
An integer vector the same size as \code{x}.
}
\description{
\code{vec_order_base()} orders vectors using \code{\link[base:order]{base::order()}}, but can handle
more complex types, like data frames and \code{\link[=vctr]{vctrs_vctr}} objects, using
vctrs principles.

\code{vec_order_base()} is mainly provided for backwards compatibility with vctrs
<= 0.3.7. New code should instead use \code{\link[=vec_order]{vec_order()}}, which has more
capabilities. The main difference between the two is that \code{vec_order()}
orders character vectors in the C locale (which is highly performance), while
\code{vec_order_base()} respects the system locale.
}
\section{Differences with \code{order()}}{

Unlike the \code{na.last} argument of \code{order()} which decides the positions of
missing values irrespective of the \code{decreasing} argument, the \code{na_value}
argument of \code{vec_order_base()} interacts with \code{direction}. If missing values
are considered the largest value, they will appear last in ascending order,
and first in descending order.
}

\section{Dependencies of \code{vec_order_base()}}{

\itemize{
\item \code{\link[=vec_proxy_order]{vec_proxy_order()}}
}
}

\examples{
x <- round(c(runif(9), NA), 3)
vec_order_base(x)
vec_order_base(x, "desc")

# Can also handle data frames
df <- data.frame(g = sample(2, 10, replace = TRUE), x = x)
vec_order_base(df)
vec_order_base(df, "desc")

# Missing values interpreted as largest values are last when
# in increasing order:
vec_order_base(c(1, NA), na_value = "largest", direction = "asc")
vec_order_base(c(1, NA), na_value = "largest", direction = "desc")
}
\keyword{internal}
