% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{vec_order_loc}
\alias{vec_order_loc}
\title{Identify ordered groups}
\usage{
vec_order_loc(x, direction = "asc", na_value = "largest")
}
\arguments{
\item{x}{A vector}

\item{direction}{Direction to sort in.
\itemize{
\item A single \code{"asc"} or \code{"desc"} for ascending or descending order
respectively.
\item For data frames, a length \code{1} or \code{ncol(x)} character vector containing
only \code{"asc"} or \code{"desc"}, specifying the direction for each column.
}}

\item{na_value}{Treatment of \code{NA} values. \code{NaN} values are treated as
equivalent to \code{NA} values.
\itemize{
\item A single \code{"largest"} or \code{"smallest"} for treating \code{NA} values as the
largest or smallest values respectively.
\item For data frames, a length \code{1} or \code{ncol(x)} character vector containing
only \code{"largest"} or \code{"smallest"}, specifying how \code{NA}s should be treated
in each column.
}}
}
\value{
A two column data frame with size equal to \code{vec_size(vec_unique(x))}.
\itemize{
\item A \code{key} column of type \code{vec_ptype(x)}.
\item A \code{loc} column of type list, with elements of type integer.
}
}
\description{
\Sexpr[results=rd, stage=render]{vctrs:::lifecycle("experimental")}

\code{vec_order_loc()} returns a data frame containing a \code{key} column with
sorted unique groups, and a \code{loc} column with the locations of each
group in \code{x}. It is similar to \code{\link[=vec_group_loc]{vec_group_loc()}}, except the groups are
returned sorted rather than by first appearance.
}
\section{Dependencies of \code{vec_order_loc()}}{

\itemize{
\item \code{\link[=vec_proxy_order]{vec_proxy_order()}}
}
}

\examples{
df <- data.frame(
  g = sample(2, 10, replace = TRUE),
  x = c(NA, sample(5, 9, replace = TRUE))
)

# `vec_order_loc()` is similar to `vec_group_loc()`, except keys are
# returned ordered rather than by first appearance.
vec_order_loc(df)

vec_group_loc(df)
}
