% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-str.R
\name{vec_print}
\alias{vec_print}
\alias{vec_print_header}
\alias{vec_print_data}
\alias{vec_print_footer}
\alias{vec_str}
\alias{vec_str_header}
\alias{vec_str_data}
\alias{vec_str_footer}
\title{\code{print() and}str()` generics.}
\usage{
vec_print(x, ...)

vec_print_header(x, ...)

vec_print_data(x, ...)

vec_print_footer(x, ...)

vec_str(x, ...)

vec_str_header(x, ...)

vec_str_data(x, ...)

vec_str_footer(x, ...)
}
\arguments{
\item{x}{A vector}

\item{...}{Additional arguments passed on to methods. See \code{\link[=print]{print()}} and
\code{\link[=str]{str()}} for commonly used options}
}
\description{
These are contructed to be more easily extensible since you can override
the \code{_header()}, \code{_data()} or \code{_footer()} components individually. The
default methods are built on top of \code{format()}.
}
\keyword{internal}
