% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{vec_proxy_equal}
\alias{vec_proxy_equal}
\title{Defunct equality proxy}
\usage{
vec_proxy_equal(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
vctrs now uses \code{\link[=vec_proxy]{vec_proxy()}}. Note that most of the time it's not
necessary to define a proxy for vector equality.
}
\keyword{internal}
