% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal.R
\name{vec_proxy_equality}
\alias{vec_proxy_equality}
\title{Equality proxy}
\usage{
vec_proxy_equality(x)
}
\arguments{
\item{x}{A vector x.}
}
\value{
A 1d atomic vector or a data frame.
}
\description{
Returns a proxy object (i.e. an atomic vector or data frame of atomic
vectors). For \link{vctr}s, this determins the behaviour of \code{==} and
\code{!=} (via \code{\link[=vec_equal]{vec_equal()}}); \code{\link[=unique]{unique()}}, \code{\link[=duplicated]{duplicated()}} (via
\code{\link[=vec_unique]{vec_unique()}} and \code{\link[=vec_duplicate_detect]{vec_duplicate_detect()}}).
}
\keyword{internal}
