% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{vec_proxy_logical}
\alias{vec_proxy_logical}
\alias{vec_restore_logical}
\title{Logical proxy}
\usage{
vec_proxy_logical(x)

vec_restore_logical(x, to)
}
\description{
\code{vec_proxy_logical()} converts an object to it's logical representation
(if available). \code{vec_restore_logical()} restores a logical proxy back to
the original type. The default methods allow any object built on a logical,
integer, or numeric, to be coerced to logical, and leave the result as is.
}
\examples{
x <- new_vctr(c(TRUE, FALSE, NA))
y <- new_vctr(c(TRUE, TRUE, TRUE))

!x
x | y
x & y
}
\keyword{internal}
