% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{vec_proxy_missing}
\alias{vec_proxy_missing}
\title{Missing proxy}
\usage{
vec_proxy_missing(x, ...)
}
\arguments{
\item{x}{A vector x.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A 1d atomic vector or a data frame.
}
\description{
Returns a proxy object (i.e. an atomic vector or data frame of atomic
vectors). For \link{vctr}s, this determines the behaviour of
\code{\link[=is.na]{is.na()}} and \code{\link[=anyNA]{anyNA()}} (via \code{\link[=vec_detect_missing]{vec_detect_missing()}}).
}
\details{
The default method calls \code{\link[=vec_proxy_equal]{vec_proxy_equal()}}, as the default
equal-able proxy should be used to detect missingness in most cases.
}
\section{Data frames}{

If the proxy for \code{x} is a data frame, the proxy function is automatically
recursively applied on all columns as well. After applying the proxy
recursively, if there are any data frame columns present in the proxy, then
they are unpacked. Finally, if the resulting data frame only has a single
column, then it is unwrapped and a vector is returned as the proxy.
}

\section{Dependencies}{

\itemize{
\item \code{\link[=vec_proxy_equal]{vec_proxy_equal()}} called by default
}
}

\keyword{internal}
