% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{vec_proxy_numeric}
\alias{vec_proxy_numeric}
\alias{vec_restore_numeric}
\title{Numeric proxy}
\usage{
vec_proxy_numeric(x)

vec_restore_numeric(x, to)
}
\description{
\code{vec_proxy_numeric()} converts an object to it's numeric representation
(if available). \code{vec_restore_numeric()} restores a numeric proxy back to
the original type. The default methods allow any object built on a logical,
integer, or numeric, to be coerced to numeric, and leave the result as is.
}
\examples{
x <- new_vctr(-10:10)

abs(x)
x ^ 2
}
\keyword{internal}
