% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{vec_proxy_order}
\alias{vec_proxy_order}
\title{Ordering proxy}
\usage{
vec_proxy_order(x, ...)
}
\arguments{
\item{x}{A vector x.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A 1d atomic vector or a data frame.
}
\description{
Returns a proxy object for ordering. For \code{\link[=vctr]{vctrs_vctr}} objects, this
determines the behavior of \code{order()} and \code{sort()} (via \code{xtfrm()}). It also
powers \code{vec_order()} and \code{vec_sort()} for other vector types.
}
\details{
The default is to call \code{\link[=vec_proxy_compare]{vec_proxy_compare()}}.

Lists are special cased, and return an integer vector that can be used to
order the elements in the list by first appearance.
}
\section{Dependencies}{

\itemize{
\item \code{\link[=vec_proxy_compare]{vec_proxy_compare()}} called by default
}
}

\examples{
# List elements are ordered by first appearance
vec_proxy_order(list(1:2, 1, 1:2, 3))
}
\keyword{internal}
