% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-tibble.R
\name{vec_ptype2.grouped_df}
\alias{vec_ptype2.grouped_df}
\title{Common prototype for grouped data frames}
\usage{
\method{vec_ptype2}{grouped_df}(x, y, ...)
}
\arguments{
\item{x}{Vector types.}

\item{y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Common prototype for grouped data frames
}
