% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptype-abbr-full.R
\name{vec_ptype_abbr}
\alias{vec_ptype_abbr}
\title{Abbreviated vector type as a string}
\usage{
vec_ptype_abbr(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A string.
}
\description{
\Sexpr[results=rd, stage=render]{vctrs:::lifecycle("maturing")}
\code{vec_ptype_abbr()} provides an abbreviated summary suitable for use in a column heading.
}
\section{S3 dispatch}{

The default method for \code{vec_ptype_abbr()} \code{\link[=abbreviate]{abbreviate()}}s \code{vec_ptype_full()}
to 8 characters. You should almost always override, aiming for 4-6
characters where possible.
}

\section{Life cycle}{

\code{vec_ptype_abbr()} is maturing, because it is the designated replacement for \code{\link[pillar:type_sum]{pillar::type_sum()}}.
}

\examples{
cat(vec_ptype_abbr(1:10))
}
\seealso{
Other vector descriptions: \code{\link{vec_ptype_full}}
}
\concept{vector descriptions}
\keyword{internal}
