% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptype-abbr-full.R
\name{vec_ptype_full}
\alias{vec_ptype_full}
\title{Vector type as a string}
\usage{
vec_ptype_full(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A string.
}
\description{
\Sexpr[results=rd, stage=render]{vctrs:::lifecycle("questioning")}
\code{vec_ptype_full()} displays the full type of the vector.
}
\section{S3 dispatch}{

The default method for \code{vec_ptype_full()} uses the first element of the
class vector. Override this method if your class has parameters that should
be prominently displayed.
}

\section{Life cycle}{

\code{vec_ptype_full()} is questioning, because it feels that more variants
of a type description need to be accounted for.
See \url{https://github.com/r-lib/vctrs/issues/139} for details.
}

\examples{
cat(vec_ptype_full(1:10))
cat(vec_ptype_full(iris))
}
\seealso{
Other vector descriptions: \code{\link{vec_ptype_abbr}}
}
\concept{vector descriptions}
\keyword{internal}
