% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle.R
\name{vec_recycle}
\alias{vec_recycle}
\title{Recycle vector types to common number of observations}
\usage{
vec_recycle(..., .nobs = NULL)
}
\arguments{
\item{...}{Vectors to recycle.}

\item{.nobs}{Number of observations}
}
\value{
A list of vectors each with equal \code{\link[=vec_obs]{vec_obs()}}, or an error
stating that recycling was not possible.
}
\description{
Recycle vector types to common number of observations
}
\examples{
# Inputs with 1 observation are recycled
vec_recycle(1:5, 5)
\dontrun{
vec_recycle(1:5, 1:2)
}

# Inputs with 0 observations
vec_recycle(1:5, integer())

# Data frames and matrices are recycled along their rows
vec_recycle(data.frame(x = 1), 1:5)
vec_recycle(array(1:2, c(1, 2)), 1:5)
vec_recycle(array(1:3, c(1, 3, 1)), 1:5)
}
