% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{vec_shape}
\alias{vec_shape}
\alias{vec_reshape}
\title{Compute and alter "shape" of a vector}
\usage{
vec_shape(..., data.frame = FALSE)

vec_reshape(x, shape)
}
\arguments{
\item{data.frame}{Does this shape represent a data frame? Data frames
must be 2 dimensional, and can only be recycled along the first dimension.}

\item{x}{For \code{shape()}, a vector giving dimensions of an object.
For \code{as_vec_shape()}, an existing data vector}
}
\description{
The shape of a vector is its \code{vec_dim()} (i.e. \code{dim()} or \code{length()})
and whether or not it's a data frame.
}
\examples{
vec_shape(1:10)
vec_shape(1:10, 1)
\dontrun{
vec_shape(1:10, 1)
}

vec_shape(mtcars)
}
\keyword{internal}
