% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{vec_slice}
\alias{vec_slice}
\alias{vec_slice<-}
\alias{vec_assign}
\title{Get or set observations in a vector}
\usage{
vec_slice(x, i, ..., error_call = current_env())

vec_slice(x, i) <- value

vec_assign(x, i, value, ..., style = "location", x_arg = "", value_arg = "")
}
\arguments{
\item{x}{A vector}

\item{i}{An index vector

For \code{vec_slice()} and \code{vec_assign(style = "location")}, an integer,
character or logical vector specifying the locations or names of the
observations to get/set. Specify \code{TRUE} to index all elements (as in
\code{x[]}), or \code{NULL}, \code{FALSE} or \code{integer()} to index none (as in \code{x[NULL]}).

For \code{vec_assign(style = "condition")}, a logical vector the same length as
\code{x} specifying the observations to set.}

\item{...}{These dots are for future extensions and must be empty.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{value}{A vector of replacement values

\code{value} is cast to the type of \code{x}, but only if they have a common type.
See below for examples of this rule.

For \code{style = "location"}, \code{value} must be size 1 or the same size as \code{i}
after \code{i} has been converted to a positive integer location vector (which
may not be the same size as \code{i} originally).

For \code{style = "condition"}, \code{value} must be size 1 or the same size as
\code{i}.}

\item{style}{The assignment style, which influences how both \code{i} and \code{value}
are interpreted (see their documentation for exact details). One of:
\itemize{
\item \code{"location"}: Most closely resembles \code{base::`[<-`}, where an
index vector is converted to a positive integer \emph{location} vector before
assignment occurs.
\item \code{"condition"}: Most closely resembles \code{\link[base:ifelse]{base::ifelse()}}, where the index
is provided as a logical \emph{condition} vector, and elements are copied from
\code{value} where the condition vector is \code{TRUE}.
}}

\item{x_arg, value_arg}{Argument names for \code{x} and \code{value}. These are used
in error messages to inform the user about the locations of
incompatible types and sizes (see \code{\link[=stop_incompatible_type]{stop_incompatible_type()}} and
\code{\link[=stop_incompatible_size]{stop_incompatible_size()}}).}
}
\value{
A vector of the same type as \code{x}.
}
\description{
This provides a common interface to extracting and modifying observations
for all vector types, regardless of dimensionality. They are analogs to \code{[}
and \verb{[<-} that match \code{\link[=vec_size]{vec_size()}} instead of \code{length()}.
}
\section{Genericity}{


Support for S3 objects depends on whether the object implements a
\code{\link[=vec_proxy]{vec_proxy()}} method.
\itemize{
\item When a \code{vec_proxy()} method exists, the proxy is sliced or assigned to and
\code{vec_restore()} is called on the result.
\item Otherwise, \code{vec_slice()} falls back to the base generic \code{[} and
\verb{vec_slice<-()} falls back to the base generic \verb{[<-}.
}

When \verb{vec_slice<-()} falls back to \verb{[<-}, it is expected that the subclass's
\verb{[<-} method can handle the following subset of cases that base R's \verb{[<-}
can also handle:
\itemize{
\item An \code{i} vector of positive integer positions (notably excluding \code{NA}).
\item A \code{value} vector of length 1 or length \code{length(i)}. If length 1, it
should be recycled by the \verb{[<-} method to the length of \code{i}.
}

If your \verb{[<-} method eventually calls base R's native \verb{[<-} code, then these
cases will be handled for you.

Note that S3 lists are treated as scalars by default, and will
cause an error if they don't implement a \code{\link[=vec_proxy]{vec_proxy()}} method.
}

\section{Differences with base R subsetting}{

\itemize{
\item \code{vec_slice()} only slices along one dimension. For
two-dimensional types, the first dimension is subsetted.
\item \code{vec_slice()} preserves attributes by default.
\item \verb{vec_slice<-()} is type-stable and always returns the same type
as the LHS.
}
}

\section{Dependencies}{

\subsection{vctrs dependencies}{
\itemize{
\item \code{\link[=vec_proxy]{vec_proxy()}}
\item \code{\link[=vec_restore]{vec_restore()}}
}
}

\subsection{base dependencies}{
\itemize{
\item \code{base::`[`}
\item \code{base::`[<-`}
}
}
}

\examples{
x <- sample(10)
x
vec_slice(x, 1:3)

# You can assign with the infix variant:
vec_slice(x, 2) <- 100
x

# Or with the regular variant that doesn't modify the original input:
y <- vec_assign(x, 3, 500)
y
x


# Slicing objects of higher dimension:
vec_slice(mtcars, 1:3)

# Type stability --------------------------------------------------

# The assign variant is type stable. It always returns the same
# type as the input.
x <- 1:5
vec_slice(x, 2) <- 20.0

# `x` is still an integer vector because the RHS was cast to the
# type of the LHS:
vec_ptype(x)

# Compare to `[<-`:
x[2] <- 20.0
vec_ptype(x)


# Note that the types must be coercible for the cast to happen.
# For instance, you can cast a double vector of whole numbers to an
# integer vector:
vec_cast(1, integer())

# But not fractional doubles:
try(vec_cast(1.5, integer()))

# For this reason you can't assign fractional values in an integer
# vector:
x <- 1:3
try(vec_slice(x, 2) <- 1.5)

# Assignment style ------------------------------------------------

# The `style` argument influences how `i` and `value` are interpreted,
# particularly with logical `i` vectors.

# `"location"` forces `i` to be converted to a positive integer location
# vector with `which(i)` before assignment occurs, and `value`'s size
# must match `i`'s AFTER that conversion is done. This matches base R's
# `[<-`, but is not always intuitive.
x <- 1:4
i <- c(TRUE, FALSE, TRUE, FALSE)

# This works because `which(i)` results in `c(1, 3)`
vec_assign(x, i, c(-1, -2))

# But this fails, even though you might intuitively think it works because
# `vec_size(i) == vec_size(value)`
value <- c(-1, -2, -3, -4)
try(vec_assign(x, i, value))

# `"condition"` assigns using that logical `i` directly, and requires that
# `value` be size 1 or the same size as the logical `i` vector. It pulls
# elements from `value` where `i` is `TRUE`, so this works:
vec_assign(x, i, value, style = "condition")
}
\keyword{internal}
