% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{vec_split_id}
\alias{vec_split_id}
\title{Locate unique groups in a vector}
\usage{
vec_split_id(x)
}
\arguments{
\item{x}{A vector to locate unique groups for.}
}
\value{
A data frame with two columns and size equal to
\code{vec_size(vec_unique(x))}. The \code{key} column has the same type as \code{x}, and
the \code{id} column is a list of integers.

Note for complex types, the default \code{data.frame} print method will be
suboptimal, and you will want to coerce into a tibble to better understand
the output.
}
\description{
This locates unique values in \code{x} and returns both the unique values and the
locations of those values. It is used to power \code{\link[=vec_split]{vec_split()}}.
}
\examples{
vec_split_id(mtcars$vs)
vec_split_id(mtcars[c("vs", "am")])

if (require("tibble")) {
  as_tibble(vec_split_id(mtcars[c("vs", "am")]))
}
}
\seealso{
\link{vec_split}
}
