% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{vec_type}
\alias{vec_type}
\alias{new_vec_type}
\alias{as_vec_type}
\alias{max.vec_type}
\title{Compute "type" of a vector}
\usage{
vec_type(x)

new_vec_type(prototype)

as_vec_type(x)

\method{max}{vec_type}(..., na.rm = FALSE)
}
\arguments{
\item{x}{For \code{vec_type()} a vector; for \code{as_vec_type()} a vector or
vectory type}
}
\description{
The type of a vector capture its essence, independent of it's values.
We record the type (including its parameters) with a 0-row subset of the
vector, which we call the prototype.
}
\details{
\code{new_vec_type()} is the low-level S3 constructor, \code{vec_type()} is a helper
for interactive exploration, and \code{as_vec_type()} for programming.
Use \code{max()} to compute the "maximum" type of a set of types; this will
return an error if no maximum type exists.
}
\examples{
vec_type(mtcars)

max(vec_type(1), vec_type(1L))
\dontrun{
max(vec_type(1), vec_type("a"))
}
}
\keyword{internal}
