% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-string.R
\name{vec_type_string}
\alias{vec_type_string}
\title{Vector type as a string}
\usage{
vec_type_string(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A string.
}
\description{
Provide a method for this generic to control how your vector type is
printed. The default method uses the first element of the class vector.
}
\examples{
cat(vec_type_string(1:10))
cat(vec_type_string(iris))
}
\keyword{internal}
