% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_of.R
\name{list_of}
\alias{list_of}
\alias{as_list_of}
\alias{new_list_of}
\alias{is_list_of}
\alias{vec_type2.list_of}
\alias{vec_cast.list_of}
\title{Construct "list_of" objects}
\usage{
list_of(..., .ptype = NULL)

as_list_of(x, ...)

new_list_of(x, .ptype)

is_list_of(x)

\method{vec_type2}{list_of}(x, y)

\method{vec_cast}{list_of}(x, to)
}
\arguments{
\item{...}{Vectors to coerce. All vectors must be 1d (i.e. no data
frames, matrices or arrays).}

\item{.ptype}{If \code{NULL}, the default, the output type is determined by
computing the common type across all elements of \code{...}.

Alternatively, you can supply \code{.ptype} to give the output known type.
If \code{getOption("vctrs.no_guessing")} is \code{TRUE} you must supply this value:
this is a convenient way to make production code demand fixed types.}

\item{x}{For \code{as_list_of()}, a vector to be coerced to list_of.}

\item{y, to}{Arguments to \code{vec_type2()} and \code{vec_cast()}.}
}
\description{
A \code{list_of} object is a list where each element has the same type.
Modifying the list with \code{$}, \code{[}, and \code{[[} preserves the constraint
by coercing all input items.
}
\examples{
x <- list_of(1:3, 5:6, 10:15)
tibble::tibble(x = x)

vec_c(list_of(1, 2), list_of(FALSE, TRUE))
}
