% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-transpose.R
\name{list_transpose}
\alias{list_transpose}
\title{Transpose a list of vectors}
\usage{
list_transpose(
  x,
  ...,
  size = NULL,
  ptype = NULL,
  x_arg = caller_arg(x),
  error_call = current_env()
)
}
\arguments{
\item{x}{A list of vectors.
\itemize{
\item Each vector must be the same size.
\item Each vector will be \link[=theory-faq-coercion]{cast} to the common type
before transposing.
\item \code{NULL} elements are not allowed.
}}

\item{...}{These dots are for future extensions and must be empty.}

\item{size}{The expected size of each element of \code{x}. If not provided,
computed as \code{vec_size(x[[1L]])}, or \code{0L} if \code{x} is empty.}

\item{ptype}{The expected type of each element of \code{x}. If not provided,
computed automatically by \code{\link[=vec_ptype_common]{vec_ptype_common()}}.}

\item{x_arg}{Argument name used in error messages.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A list of vectors with the following invariants:

For the list:
\itemize{
\item \verb{vec_ptype(list_transpose(x)) == <list>}
\item \code{vec_size(list_transpose(x)) == size \%||\% vec_size(x[[1L]]) \%||\% 0L}
}

For the list elements:
\itemize{
\item \code{vec_ptype(list_transpose(x)[[i]]) == vec_ptype_common(!!!x, .ptype = ptype)}
\item \code{vec_size(list_transpose(x)[[i]]) == vec_size(x)}
}
}
\description{
\code{list_transpose()} takes a list of vectors of equal size, transposes it, and
returns a new list of vectors of equal size.

To predict the output from \code{list_transpose()}, swap the size of the list
with the size of the list elements. For example:
\itemize{
\item Input: List of size 2, elements of size 3
\item Output: List of size 3, elements of size 2
}
}
\details{
In an ideal world, this function would transpose a data frame. Data frames
have a few desirable properties:
\itemize{
\item Each column is the same size
\item \code{NULL} columns are not allowed
}

The downside is that data frames must have names. When transposing, names
are meaningless, both on the input and output.

The compromise struck here is to allow a list, which doesn't require names,
but to enforce that each element of the list must be the same size and can't
be \code{NULL}.
}
\examples{
# Input:
# - List size 3
# - Element size 2
# Output:
# - List size 2
# - Element size 3
list_transpose(list(1:2, 3:4, 5:6))

# With data frames
x <- data_frame(a = 1:2, b = letters[1:2])
y <- data_frame(a = 3:4, b = letters[3:4])
list_transpose(list(x, y))

# With size 0 elements, the invariants are a bit tricky!
# This must return a size 0 list, but then you lose expected
# type (integer) and size (2) information about the elements.
# Losing that information makes it difficult to reverse the
# transposition.
#
# Input:
# - List size 2
# - Element size 0
# Output:
# - List size 0
# - Element size 2
x <- list(integer(), integer())
out <- list_transpose(x)
out

# Note how transposing a second time doesn't recover the original list
list_transpose(out)

# To work around this, provide the lost `size` and `ptype` manually
list_transpose(out, size = vec_size(x), ptype = vec_ptype_common(!!!x))

# Note that this function doesn't recycle elements. This is purposeful,
# it is meant for transposing "rectangular lists", which are lists with
# elements of equal size.
x <- list(1, 2:3)
try(list_transpose(x))

# If you'd like to pad with a missing value rather than erroring,
# you might do something like this, which left-pads
x <- list(1, 2:5, 6:7)
try(list_transpose(x))

sizes <- list_sizes(x)
size <- max(sizes)
index <- which(sizes != size)
x[index] <- lapply(
  index,
  function(i) vec_c(rep(NA, times = size - sizes[[i]]), x[[i]])
)
list_transpose(x)

# `NULL` values aren't allowed in `list_transpose()` because theoretically
# this function is meant to transpose a data frame, which can't have `NULL`
# columns.
x <- list(1:3, NULL, 5:7, NULL)
try(list_transpose(x))

# Either drop the `NULL` values or replace them with something else
list_transpose(list_drop_empty(x))

na <- list(vec_rep(NA, vec_size_common(!!!x)))
x <- vec_assign(x, vec_detect_missing(x), na)
list_transpose(x)
}
