% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-rcrd.R
\name{new_rcrd}
\alias{new_rcrd}
\alias{ses}
\alias{rcrd}
\title{rcrd (record) S3 class}
\usage{
new_rcrd(fields, ..., class = character())
}
\arguments{
\item{fields}{A list or a data frame. Lists must be rectangular
(same sizes), and contain uniquely named vectors (at least
one). \code{fields} is validated with \code{\link[=df_list]{df_list()}} to ensure uniquely
named vectors.}

\item{...}{Additional attributes}

\item{class}{Name of subclass.}
}
\description{
The rcrd class extends \link{vctr}. A rcrd is composed of 1 or more \link{field}s,
which must be vectors of the same length. Is designed specifically for
classes that can naturally be decomposed into multiple vectors of the same
length, like \link{POSIXlt}, but where the organisation should be considered
an implementation detail invisible to the user (unlike a \link{data.frame}).
}
\details{
Record-style objects created with \code{\link[=new_rcrd]{new_rcrd()}} do not do much on their own.
For instance they do not have a default \code{\link[=format]{format()}} method, which means
printing the object causes an error. See the \emph{Record-style objects} section
from \code{vignette("s3-vector")} for details on implementing methods for
record vectors.
}
\keyword{internal}
