% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr.R
\name{new_vctr}
\alias{new_vctr}
\alias{vctr}
\alias{vec_print_header}
\alias{vec_print_data}
\alias{vec_print_footer}
\title{vctr (vector) S3 class}
\usage{
new_vctr(.data, ..., class = character())

vec_print_header(x, ...)

vec_print_data(x, ...)

vec_print_footer(x, ...)
}
\arguments{
\item{...}{Name-value pairs defining attributes}

\item{class}{Name of subclass.}

\item{x}{Foundation of class. Must be a vector}
}
\description{
This abstract class provides a set of useful default methods that makes it
considerably easier to get started with a new S3 vector class. See
\code{vignette("S3-vector")} to learn how to use it to create your own S3
vector classes.
}
\section{Base methods}{

The vctr class provides methods for many base generics using a smaller
set of generics defined by this package.
\itemize{
\item \code{[[} and \code{[} use \code{NextMethod()} dispatch to the underlying base function,
then restore attributes with \code{vec_restore()}.
\code{rep()} and \code{length<-} work similarly.
\item \code{[[<-} and \code{[<-} cast \code{value} to same type as \code{x}, then call
\code{NextMethod()}.
\item \code{==}, \code{!=}, \code{unique()}, any \code{duplicated()} use \link{vec_proxy_equal}.
\item \code{<}, \code{<=}, \code{>=}, \code{>}, \code{min()}, \code{max()}, \code{median()}, \code{quantile()},
and \code{xtfrm()} methods use \link{vec_proxy_compare}.
\item \code{!}, \code{&}, and \code{|} methods use \code{\link[=vec_proxy_logical]{vec_proxy_logical()}}.
\item Mathematical operations use \code{\link[=vec_proxy_numeric]{vec_proxy_numeric()}}.
\item \code{as.logical()}, \code{as.integer()}, \code{as.numeric()}, \code{as.character()},
\code{as.Date()} and \code{as.POSIXct()} methods call \code{vec_cast()}.
The \code{as.list()} method calls \code{[[} repeatedly, and the \code{as.data.frame()}
method uses a standard technique to wrap a vector in a data frame.
\item \code{dims()}, \code{dims<-()}, \code{dimnames()}, \code{dimnames<-}, \code{levels()}, and
\code{levels<-} methods throw errors.
}
}

\keyword{internal}
