% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-factor.R
\name{partial_factor}
\alias{partial_factor}
\title{Partially specify a factor}
\usage{
partial_factor(levels = character())
}
\arguments{
\item{levels}{Character vector of labels.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This special class can be passed as a \code{ptype} in order to specify that the
result should be a factor that contains at least the specified levels.
}
\examples{
pf <- partial_factor(levels = c("x", "y"))
pf

vec_ptype_common(factor("v"), factor("w"), .ptype = pf)

}
\keyword{internal}
