% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unknown.R
\name{unknown}
\alias{unknown}
\alias{is_unknown}
\alias{vec_type2.unknown}
\title{A vector of unknown type}
\usage{
unknown(n = 0)

is_unknown(x)

\method{vec_type2}{unknown}(x, y)
}
\arguments{
\item{n}{Length of vector}
}
\description{
This is a special vector type used to represent of known length but
unknown type.
}
\examples{
vec_ptype()
vec_ptype(NA)

vec_ptype(unknown(), "x")
}
\keyword{internal}
