% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-unspecified.R
\name{unspecified}
\alias{unspecified}
\title{A 1d vector of unspecified type}
\usage{
unspecified(n = 0)
}
\arguments{
\item{n}{Length of vector}
}
\description{
This is a \link[=new_partial]{partial type} used to represent logical vectors
that only contain \code{NA}. These require special handling because we want to
allow \code{NA} to specify missingness without requiring a type.
}
\examples{
vec_ptype_show()
vec_ptype_show(NA)

vec_c(NA, factor("x"))
vec_c(NA, Sys.Date())
vec_c(NA, Sys.time())
vec_c(NA, list(1:3, 4:5))
}
\keyword{internal}
