% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{vctrs-conditions}
\alias{vctrs-conditions}
\alias{stop_incompatible_type}
\alias{stop_incompatible_cast}
\alias{stop_incompatible_op}
\alias{stop_lossy_cast}
\alias{suppress_errors_lossy_cast}
\title{Custom conditions for vctrs package}
\usage{
stop_incompatible_type(x, y, details = NULL, ..., message = NULL,
  .subclass = NULL)

stop_incompatible_cast(x, y, details = NULL, ..., message = NULL,
  .subclass = NULL)

stop_incompatible_op(op, x, y, details = NULL, ..., message = NULL,
  .subclass = NULL)

stop_lossy_cast(x, to, result, locations = NULL, details = NULL, ...,
  message = NULL, .subclass = NULL)

suppress_errors_lossy_cast(expr, x_ptype = NULL, to_ptype = NULL)
}
\arguments{
\item{x, y}{Vectors}

\item{details}{Any additional human readable details}

\item{..., message, .subclass}{Only use these fields when creating a subclass.}

\item{to}{Type to cast to.}

\item{result}{The result of a lossy cast.}

\item{locations}{For \code{stop_lossy_cast()}, an optional vector giving the
locations where \code{x} lost information.}

\item{x_ptype, to_ptype}{Suppress only the casting errors where \code{x}
or \code{to} match these \link[=vec_type]{prototypes}.}

\item{subclass}{Use if you want to further customise the class}
}
\description{
These errors and warnings have custom classes and structures to make
testing easier.

By default, lossy casts are an error. Use
\code{suppress_errors_lossy_cast()} to silence these errors and continue
with the partial results. The lost values are typically set to \code{NA}.
}
\keyword{internal}
