% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-package.R
\docType{package}
\name{vctrs-package}
\alias{vctrs}
\alias{vctrs-package}
\title{vctrs: Vector Helpers}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}
\Sexpr[results=rd, stage=render]{vctrs:::lifecycle("maturing")}

Defines new notions of prototype and size that are
used to provide tools for consistent and well-founded type-coercion
and size-recycling, and are in turn connected to ideas of type- and
size-stability useful for analysing function interfaces.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/vctrs}
  \item Report bugs at \url{https://github.com/r-lib/vctrs/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Authors:
\itemize{
  \item Lionel Henry \email{lionel@rstudio.com}
  \item Davis Vaughan \email{davis@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
