% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-package.R
\docType{package}
\name{vctrs-package}
\alias{vctrs}
\alias{vctrs-package}
\title{vctrs: Vector Helpers}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Defines new notions of prototype and size that are
used to provide tools for consistent and well-founded type-coercion
and size-recycling, and are in turn connected to ideas of type- and
size-stability useful for analysing function interfaces.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vctrs.r-lib.org/}
  \item Report bugs at \url{https://github.com/r-lib/vctrs/issues}
}

}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@posit.co}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co}
  \item Davis Vaughan \email{davis@posit.co}
}

Other contributors:
\itemize{
  \item data.table team (Radix sort based on data.table's forder() and their contribution to R's order()) [copyright holder]
  \item Posit, PBC [copyright holder, funder]
}

}
\keyword{internal}
