% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscript.R
\name{vec_as_subscript}
\alias{vec_as_subscript}
\alias{vec_as_subscript2}
\title{Convert to a base subscript type}
\usage{
vec_as_subscript(
  i,
  ...,
  allow_types = c("indicator", "location", "name"),
  arg = "i"
)

vec_as_subscript2(i, ..., allow_types = c("location", "name"), arg = "i")
}
\arguments{
\item{i}{An integer, character or logical vector specifying the
locations or names of the observations to get/set. Specify
\code{TRUE} to index all elements (as in \code{x[]}), or \code{NULL}, \code{FALSE} or
\code{integer()} to index none (as in \code{x[NULL]}).}

\item{allow_types}{Experimental. Character vector indicating one or
several types of location to be allowed as input: \code{"indicator"},
\code{"location"}, or \code{"name"}. Indicators must be subtypes of
logical. Locations must be coercible to integers (possibly from
a larger type like double). Names must be subtypes of character.}

\item{arg}{The argument name to be displayed in error messages when
\code{vec_as_location()} and \code{vec_as_location2()} are used to check the
type of a function input.}
}
\description{
\Sexpr[results=rd, stage=render]{vctrs:::lifecycle("experimental")}

Convert \code{i} to the base type expected by \code{\link[=vec_as_location]{vec_as_location()}} or
\code{\link[=vec_as_location2]{vec_as_location2()}}. The values of the subscript type are
not checked in any way (length, missingness, negative elements).
}
\keyword{internal}
