% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{vec_assert}
\alias{vec_assert}
\title{Assert an argument has known prototype and/or size}
\usage{
vec_assert(x, ptype = NULL, size = NULL)
}
\arguments{
\item{x}{A vector argument to check.}

\item{ptype}{Prototype to compare against.}

\item{size}{Size to compare against}
}
\value{
Either an error, or \code{x}, invisibly.
}
\description{
Assert an argument has known prototype and/or size
}
