% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{vec_cbind_frame_ptype}
\alias{vec_cbind_frame_ptype}
\title{Frame prototype}
\usage{
vec_cbind_frame_ptype(x, ...)
}
\arguments{
\item{x}{A data frame.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is an experimental generic that returns zero-columns variants
of a data frame. It is needed for \code{\link[=vec_cbind]{vec_cbind()}}, to work around the
lack of colwise primitives in vctrs. Expect changes.
}
\keyword{internal}
