% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{vec_coerce}
\alias{vec_coerce}
\title{Coerce vectors to shared type}
\usage{
vec_coerce(..., .type = NULL)
}
\arguments{
\item{...}{Vectors to coerce.}

\item{.type}{If \code{NULL}, the default, the output type is determined by
computing the common type across all inputs.

Alternatively, you can supply \code{.type} to force the output to have known
type, or to die trying. \code{.type = character()} and \code{.type = list()} will
succeed for all vectors. See \code{\link[=vec_cast]{vec_cast()}} for more details.}
}
\value{
A \link{list_of} input vectors coerced to \code{.type}, an error stating
that a common type could not be found, or an error stating that casting
a input to \code{.type} was not possible.
}
\description{
This function is a useful as a building block for other functions. The
implementation also illustrates the key ideas of the vctrs type system.
}
\examples{
vec_coerce(factor("a"), factor(c("a", "b")))
vec_coerce(factor("a"), Sys.Date(), .type = list())
}
