% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{vec_data}
\alias{vec_data}
\title{Return the underlying data from a vector}
\usage{
vec_data(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The data underlying \code{x}, free from an attributes.
}
\description{
Currently, due to the underlying memory architecture of vectors, this
requires a full copy of the underlying data.
}
