% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{vec_data}
\alias{vec_data}
\title{Extract underlying data}
\usage{
vec_data(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The data underlying \code{x}, free from an attributes.
}
\description{
Extract the data underlying an S3 vector object, i.e. the underlying
atomic vector or list. Currently, due to the underlying memory architecture
of R, this creates a full copy of the underlying data.
}
\seealso{
See \code{\link[=vec_restore]{vec_restore()}} for the inverse operation: it restores
attributes given a bare vector and a prototype;
\code{vec_restore(vec_data(x), x)} will always yield \code{x}.
}
