% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{vec_deduplicate}
\alias{vec_deduplicate}
\title{Modify a function to act on a deduplicated vector input}
\usage{
vec_deduplicate(f)
}
\arguments{
\item{f}{Function whose first argument will be deduplicated.}
}
\value{
A deduplicated function
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The deduplicated function acts on the unique values in the first input \code{x}
and expands the output back to return. The return value is equivalent to \code{f(x)}
but is significantly faster for inputs with significant duplication.
}
\examples{
x <- sample(LETTERS, 10)
x

large_x <- sample(rep(x, 10))
length(large_x)

long_func <- function(x) for(i in x) {Sys.sleep(0.001)}

system.time(y <- long_func(large_x))
system.time(y2 <- vec_deduplicate(long_func)(large_x))
all(y == y2)
}
