% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-deprecated.R
\name{vec_empty}
\alias{vec_empty}
\title{Is a vector empty}
\usage{
vec_empty(x)
}
\arguments{
\item{x}{An object.}
}
\description{
\ifelse{html}{\figure{lifecycle-defunct.svg}{options: alt='Defunct lifecycle'}}{\strong{Defunct}}

This function is defunct, please use \code{\link[=vec_is_empty]{vec_is_empty()}}.
}
\keyword{internal}
