% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{vec_first_loc}
\alias{vec_first_loc}
\title{Locate first occurrences}
\usage{
vec_first_loc(x)
}
\arguments{
\item{x}{A vector (including a data frame).}
}
\value{
An integer vector the same size as \code{x}.
}
\description{
\code{vec_first_loc()} returns an integer vector giving, for each element of \code{x},
the location of its first occurrence.
}
\section{Missing values}{

In most cases, missing values are not considered to be equal, i.e.
\code{NA == NA} is not \code{TRUE}. This behaviour would be unappealing here,
so these functions consider all \code{NAs} to be equal. (Similarly,
all \code{NaN} are also considered to be equal.)
}

\examples{
x <- c(10, 10, 20, 30, 30, 40)

vec_first_loc(x)

vec_first_loc(c(NA, NA, NaN, NaN))
}
