% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-deprecated.R
\name{vec_is_list}
\alias{vec_is_list}
\title{List checks}
\usage{
vec_is_list(x)
}
\arguments{
\item{x}{For \verb{vec_*()} functions, an object. For \verb{list_*()} functions, a
list.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions have been deprecated as of vctrs 0.6.0.
\itemize{
\item \code{vec_is_list()} has been renamed to \code{\link[=obj_is_list]{obj_is_list()}}.
\item \code{vec_check_list()} has been renamed to \code{\link[=obj_check_list]{obj_check_list()}}.
}
}
\keyword{internal}
