% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{vec_math}
\alias{vec_math}
\alias{vec_math_base}
\title{Mathematical operations}
\usage{
vec_math(fun, x, ...)

vec_math_base(fun, x, ...)
}
\arguments{
\item{fun}{An mathematical function as a string}

\item{x}{A vector}

\item{...}{An additional arguments.}
}
\description{
This generic provides a common dispatch mechanism for all regular unary
mathematical functions (see \code{\link[=vec_arith]{vec_arith()}} infix functions). It is used
as a common wrapper around the Summary group generics, the Math group
generics, and a hanful of other mathematical functions like \code{mean()}.
}
\examples{
x <- new_vctr(c(1, 2.5, 10))
x

abs(x)
sum(x)
cumsum(x)
}
\keyword{internal}
