% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{vec_order}
\alias{vec_order}
\alias{vec_sort}
\title{Order and sort vectors}
\usage{
vec_order(x, direction = "asc", na_value = "largest")

vec_sort(x, direction = "asc", na_value = "largest")
}
\arguments{
\item{x}{A vector}

\item{direction}{Direction to sort in. Defaults to \code{asc}ending.}

\item{na_value}{Should \code{NA}s be treated as the largest or smallest values?}
}
\value{
\itemize{
\item \code{vec_order()} an integer vector the same size as \code{x}.
\item \code{vec_sort()} a vector with the same size and type as \code{x}.
}
}
\description{
Order and sort vectors
}
\section{Dependencies of \code{vec_order()}}{

\itemize{
\item \code{\link[=vec_proxy_compare]{vec_proxy_compare()}}
}
}

\section{Dependencies of \code{vec_sort()}}{

\itemize{
\item \code{\link[=vec_proxy_compare]{vec_proxy_compare()}}
\item \code{\link[=vec_order]{vec_order()}}
\item \code{\link[=vec_slice]{vec_slice()}}
}
}

\examples{
x <- round(c(runif(9), NA), 3)
vec_order(x)
vec_sort(x)
vec_sort(x, "desc")

# Can also handle data frames
df <- data.frame(g = sample(2, 10, replace = TRUE), x = x)
vec_order(df)
vec_sort(df)
vec_sort(df, "desc")
}
