% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{vec_order}
\alias{vec_order}
\alias{vec_sort}
\title{Order and sort vectors}
\usage{
vec_order(x, direction = "asc", na_value = c("largest", "smallest"))

vec_sort(x, direction = "asc", na_value = c("largest", "smallest"))
}
\arguments{
\item{x}{A vector}

\item{direction}{Direction to sort in.
\itemize{
\item Unless \code{x} is a data frame, this should be a single \code{"asc"} or
\code{"desc"} for ascending or descending order respectively.
\item For data frames, this is allowed to be a vector of \code{"asc"} and \code{"desc"}
with length equal to the number of columns in \code{x} specifying the
direction to order each column in.
}}

\item{na_value}{Should \code{NA}s be treated as the largest or smallest values?}
}
\value{
\itemize{
\item \code{vec_order()} an integer vector the same size as \code{x}.
\item \code{vec_sort()} a vector with the same size and type as \code{x}.
}
}
\description{
\code{vec_order()} computes the order of \code{x}. For data frames, the order is
computed along the rows by computing the order of the first column and
using subsequent columns to break ties.

\code{vec_sort()} sorts \code{x} by computing its order and using \code{vec_slice()} to
rearrange.
}
\details{
Character vectors are ordered in the C-locale. This is different from
\code{base::order()}, which respects \code{base::Sys.setlocale()}, but should
produce more reproducible results between different sessions and platforms.
One immediate difference is that ordering is case-sensitive in the C-locale.
Sorting \code{c("b", "C", "a")} with \code{vec_sort()} will return \code{c("C", "a", "b")},
but with \code{base::order()} will return \code{c("a", "b", "C")} unless
\code{base::order(method = "radix")} is explicitly set, which also uses the
C-locale. That said, this typically does not introduce any practical issues.

Character vectors are always sorted in UTF-8. If any non-UTF-8
characters are detected, they are translated to UTF-8 first.
}
\section{Dependencies of \code{vec_order()}}{

\itemize{
\item \code{\link[=vec_proxy_order]{vec_proxy_order()}}
}
}

\section{Dependencies of \code{vec_sort()}}{

\itemize{
\item \code{\link[=vec_proxy_order]{vec_proxy_order()}}
\item \code{\link[=vec_order]{vec_order()}}
\item \code{\link[=vec_slice]{vec_slice()}}
}
}

\examples{
x <- round(c(runif(9), NA), 3)
vec_order(x)
vec_sort(x)
vec_sort(x, "desc")

# Can also handle data frames
df <- data.frame(g = sample(2, 10, replace = TRUE), x = x)
vec_order(df)
vec_sort(df)
vec_sort(df, "desc")
}
