% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{vec_proxy}
\alias{vec_proxy}
\alias{vec_proxy_equality}
\alias{vec_proxy_order}
\title{Equality and ordering proxies}
\usage{
vec_proxy_equality(x)

vec_proxy_order(x)
}
\arguments{
\item{x}{A vector x.}
}
\value{
A 1d atomic vector, 1d list, or a data frame. For these functions,
\code{NA}s (and \code{NaN}s) are considered to be equal, and smaller than all
other values (i.e. they're equivalent to \code{-Inf}).
}
\description{
\code{vec_proxy_equality()} powers \code{==} and the hashing functions (and which
powers \code{vec_duplicated()} etc).
\code{vec_proxy_order()} powers \code{order()}, \code{sort()}, \code{<}, \code{>}, and so on.
Override these methods if your class has different equality or ordering
semantics.
}
\keyword{internal}
