% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{vec_proxy}
\alias{vec_proxy}
\alias{vec_proxy_equality}
\alias{vec_proxy_compare}
\title{Equality and ordering proxies}
\usage{
vec_proxy_equality(x)

vec_proxy_compare(x)
}
\arguments{
\item{x}{A vector x.}
}
\value{
A 1d atomic vector or a data frame.
}
\description{
\itemize{
\item \code{vec_proxy_equality()} powers \code{==} and the hashing functions
(and which #' powers \code{vec_duplicated()} etc).
\item \code{vec_proxy_compare()} powers \code{\link[=order]{order()}} and \code{\link[=sort]{sort()}} (via \code{\link[=xtfrm]{xtfrm()}});
\code{<}, \code{>}, \code{>=} and \code{<=}; and \code{\link[=min]{min()}}, \code{\link[=max]{max()}}, \code{\link[=median]{median()}}, and
\code{\link[=quantile]{quantile()}}.
}

Override these methods if your class has different equality or ordering
semantics.
}
\keyword{internal}
