% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{vec_proxy_compare}
\alias{vec_proxy_compare}
\title{Comparison proxy}
\usage{
vec_proxy_compare(x)
}
\arguments{
\item{x}{A vector x.}
}
\value{
A 1d atomic vector or a data frame.
}
\description{
Returns a proxy object (i.e. an atomic vector or data frame of atomic
vectors). For \link{vctr}s, this determins the behaviour of \code{\link[=order]{order()}} and
\code{\link[=sort]{sort()}} (via \code{\link[=xtfrm]{xtfrm()}}); \code{<}, \code{>}, \code{>=} and \code{<=} (via \code{\link[=vec_compare]{vec_compare()}});
and \code{\link[=min]{min()}}, \code{\link[=max]{max()}}, \code{\link[=median]{median()}}, and \code{\link[=quantile]{quantile()}}.
}
\details{
The default method assumes that all classes built on top of atomic
vectors or records are orderable. If your class is not, you will need
to provide a \code{vec_proxy_compare()} method that throws an error. Note
that the default \code{\link[=vec_proxy_equal]{vec_proxy_equal()}} method calls \code{vec_proxy_compare()} so
if your object is equal-able but not comparable, you'll need to provide
methods for both generics.
}
\keyword{internal}
