% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-tibble.R
\name{vec_ptype2.grouped_df}
\alias{vec_ptype2.grouped_df}
\alias{vec_cast.grouped_df}
\title{Double dispatch methods for grouped data frames}
\usage{
\method{vec_ptype2}{grouped_df}(x, y, ...)

\method{vec_cast}{grouped_df}(x, to, ...)
}
\arguments{
\item{x}{Vector types.}

\item{y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}
}
\description{
Double dispatch methods for grouped data frames
}
