% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-deprecated.R
\name{vec_repeat}
\alias{vec_repeat}
\title{Expand the length of a vector}
\usage{
vec_repeat(x, each = 1L, times = 1L)
}
\arguments{
\item{x}{A vector.}

\item{each}{Number of times to repeat each element of \code{x}.}

\item{times}{Number of times to repeat the whole vector of \code{x}.}
}
\value{
A vector the same type as \code{x} with size \code{vec_size(x) * times * each}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{vec_repeat()} has been replaced with \code{\link[=vec_rep]{vec_rep()}} and \code{\link[=vec_rep_each]{vec_rep_each()}} and is
deprecated as of vctrs 0.3.0.
}
\keyword{internal}
