% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-deprecated.R
\name{vec_type}
\alias{vec_type}
\alias{vec_type_common}
\alias{vec_type2}
\title{Deprecated type functions}
\usage{
vec_type(x)

vec_type_common(..., .ptype = NULL)

vec_type2(x, y, ...)
}
\arguments{
\item{x, y, ..., .ptype}{Arguments for deprecated functions.}
}
\description{
\ifelse{html}{\figure{lifecycle-deprecated.svg}{options: alt='Deprecated lifecycle'}}{\strong{Deprecated}}

These functions have been renamed:
\itemize{
\item \code{vec_type()} => \code{\link[=vec_ptype]{vec_ptype()}}
\item \code{vec_type2()} => \code{\link[=vec_ptype2]{vec_ptype2()}}
\item \code{vec_type_common()} => \code{\link[=vec_ptype_common]{vec_ptype_common()}}
}
}
\keyword{internal}
